GEOS-3830) UpdateSequence is not getting updated
git-svn-id: http://svn.codehaus.org/geoserver/branches/2.0.x@14063 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: c13f6a85238ebff787cdc7ade67aec18842dde99 [formerly 6314efdd05e5b6fce5ade3b2243acb2b71830ae6]
Former-commit-id: 015a03b570e4ca72ddc5d1d137d5c8d73bff7c58

+++ src/main/src/main/java/org/geoserver/config/GeoServerInfo.java
@@ -187,12 +187,12 @@ public interface GeoServerInfo extends Info {
      * document.
      * </p>
      */
-    int getUpdateSequence();
+    long getUpdateSequence();
     
     /**
      * Sets the update sequence.
      */
-    void setUpdateSequence( int updateSequence );
+    void setUpdateSequence( long updateSequence );
     
     /**
      * The size of the cache for feature type objects.
